fix tag iter, organize types
