+++ noita_api/src/noita/types/component_data.rs
@@ -1,0 +1,1 @@
+
