+++ noita_api/src/noita/types/objects.rs
@@ -1,0 +1,11 @@
+#[repr(C)]
+#[derive(Debug)]
+pub struct ExplosionConfig {
+    //TODO find some data maybe
+}
+
+#[repr(C)]
+#[derive(Debug)]
+pub struct GridCosmeticParticleConfig {
+    //TODO find some data maybe
+}
